soroban-rpc: Decouple simulateTransaction method and preflight computation
